CLI Helper methods extraction
